+++ toolbox/start.py
@@ -10,7 +10,7 @@ if __name__ == "__main__":
     envFile = pathlib.Path(dotenv_file)
     os.system("clear")
     loaderIntro()
-    load_dotenv(dotenv_file)
+    loadVarFile()
     if environ.get('FIRST_RUN') == "1":
         #first run stuff
         print("* This is the first time you've launched start.py, loading config menus.")
@@ -19,13 +19,13 @@ if __name__ == "__main__":
         dotenv.set_key(dotenv_file, "SETUP_STATUS", "2")
         dotenv.set_key(dotenv_file, "EASY_VERSION", easyVersion)
         isFirstRun(dotenv_file)
-        load_dotenv(dotenv_file)
+        loadVarFile()
         if environ.get('SETUP_STATUS') == "0":
             getShardMenu(dotenv_file)
             getNodeType(dotenv_file)
             setMainOrTest(dotenv_file)
             getExpressStatus(dotenv_file)
-            load_dotenv(dotenv_file)
+            loadVarFile()
             checkForInstall()
             setAPIPaths(dotenv_file)
             dotenv.unset_key(dotenv_file, "FIRST_RUN")
@@ -42,7 +42,7 @@ if __name__ == "__main__":
         getShardMenu(dotenv_file)
         getNodeType(dotenv_file)
         setMainOrTest(dotenv_file)
-        load_dotenv(dotenv_file)
+        loadVarFile()
         setAPIPaths(dotenv_file)
         nodeType = environ.get("NODE_TYPE")
         if nodeType == "regular":
